rename /validate --> /types
I realized all these functions were type-checking functions not validation functions
