+++ src/utilities/types/isAbbreviation.js
