+++ src/utilities/types/isGlottoCode.js
